ADDED confirm dialog

+++ web-ui/src/app/modules/person/components/person-detail/person-detail.component.ts
@@ -1,10 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { of } from 'rxjs';
+
 import { Person } from 'src/app/modules/person/person.model';
 import { PersonApi } from 'src/app/modules/person/person.api';
 import { UiUrls } from 'src/app/utils/urls';
 import { ErrorService } from 'src/app/modules/shared/services/error.service';
+import { DialogService } from 'src/app/modules/shared/dialogs/dialog.service';
+import { ConfirmDialogData } from 'src/app/modules/shared/dialogs/confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-person-detail',
@@ -19,7 +22,8 @@ export class PersonDetailComponent implements OnInit {
     private readonly route: ActivatedRoute,
     private readonly router: Router,
     private readonly api: PersonApi,
-    private readonly err: ErrorService
+    private readonly err: ErrorService,
+    private readonly dialog: DialogService
   ) {}
 
   ngOnInit(): void {
@@ -28,11 +32,21 @@ export class PersonDetailComponent implements OnInit {
   }
 
   delete(id: string) {
-    this.api.delete(id).subscribe({
-      next: () => {
-        this.router.navigate([UiUrls.person.list()]);
-      },
-      error: this.err.handleHttpError('delete person', new Person()),
-    });
+    this.dialog
+      .confirm(
+        new ConfirmDialogData(
+          'confirm',
+          'Are you sure you want to delete this person?'
+        )
+      )
+      .subscribe((yes) => {
+        if (yes)
+          this.api.delete(id).subscribe({
+            next: () => {
+              this.router.navigate([UiUrls.person.list()]);
+            },
+            error: this.err.handleHttpError('delete person', new Person()),
+          });
+      });
   }
 }
